+++ src/main/java/otus/dto/store/StoreResponseDTO.java
@@ -6,10 +6,9 @@ import lombok.Data;
 public class StoreResponseDTO {
   
   private Boolean complete;
-  private int id;
-  private int petId;
+  private long id;
+  private long petId;
   private int quantity;
   private String shipDate;
   private String status;
-  
 }
